backend/requirements.txt
numpy==1.21.6
opencv-python==4.11.0.86
packaging==24.0
pdf2image==1.17.0
Pillow==9.5.0
pytesseract==0.3.10
